Fixed issue with not correct status updates. Reorganized components
