+++ frontend/ohio-house-scraper-app/src/components/Body.module.css
@@ -1,5 +1,12 @@
 .tools {
     width: 100%;
+    height: 10%;
+    padding: .5rem 0.5rem;
+
+    background-color: rgb(207, 207, 207);
+
+    display: flex;
+    align-items: center;
 }
 
 .tools > div > button {
@@ -57,18 +64,8 @@
     flex-direction: column;
 }
 
-.tools {
-    width: 100%;
-    height: 10%;
-    padding: .5rem 0.5rem;
 
-    background-color: rgb(207, 207, 207);
-
-    display: flex;
-    align-items: center;
-}
-
-.outputContainer {
+.repListing {
     width: 100%;
     height: 100%;
     padding: 0.5rem;
